requirements.txt
Flask
google-generativeai
fpdf